Locking Jekyll in 3.3.0 and sitemap in 0.12.1
I am currently doing a revision in the translation and I still receive
error messages from jekyll-sitemap as in the attached image, I solved the
problem by locking the version of Jekyll in 3.3 so I could use the plugin
for sitemap in version 0.12.1.

I would suggest to test the site in the versions suggested in the link:
https://pages.github.com/versions/,

I am already running 3.3.0 with no problems, but I am still making some
adjustments since with the uOS site I am having a Jekyll crash at startup.