new sidebar tralation aproach.
Previously the sidebar had code that selected each text to be presented,
becoming a large file full of `if` instructions for each possible
language to be inserted.

In the new approach the file is selected according to the language,
being used the default when not informed or when it is the English
language.