Improvements in the new approach to posting posts and articles.
The advantage of this approach is the automatic listing on the
respective page when adding new articles. In using the previous model
you had to manually add the list.

Articles that were previously treated as pages are now treated as post,
to differentiate them, the "post" and "articles" layout should be used
respectively.

The old folder "/ page / articles" now only contains the file that
generates an index for listing articles published with the layout of the
same name. The files that were there were moved to the folder "_post /
articles" which are now adjusted for use in this format of posts but
with the aforementioned remarks.

To prevent articles from being listed in the default paigning and on the
/ blog / (old / news /) page, they must be marked with the hidden
parameter that is observed by the pager to prevent publications that
have it as true non-matching.

All other posts should be made as default using the "post" layout.

The files that were in the folder "_post / articles" were moved to
"_post / news" and had the "article" category removed to be used only in
posts that fall into this category.

The file that generates the index page of the blog is now in / pages /
blog to keep the same pattern of page creation.

An algorithm has been added that prevents the news frame from being
displayed when displaying the post listings on the link / blog /.

The parameters in the _config.yml file that change the number of posts
(news) listed have been changed so that you can group by layout used.

Layout articles were created which is nothing more than a default
wrapper.

Also added were two icons with flags of the current languages ​​for
future use and other small changes.